tile_server/modis_pipeline/requirements.txt
beautifulsoup4==4.13.3
boto3==1.37.26
botocore==1.37.26
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
jmespath==1.0.1
pyhdf==0.11.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
